+++ src/fairscoring/plots/__init__.py
@@ -1,0 +1,7 @@
+"""
+Plots to visualize and analyze different bias metrics
+"""
+from ._calibration import plot_groupwise_score_calibration
+from ._integral import plot_groupwise_cdfs, plot_cdf_diffs
+
+__all__ = ['plot_groupwise_score_calibration', 'plot_groupwise_cdfs', 'plot_cdf_diffs']