+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Change Log
 
+## [0.2.0] - 20??-??-??
+
+### Added
+#### Plots
+- Calibration plot `fairscoring.plots.plot_groupwise_score_calibration` to visualize calibration bias.
+- Cumulative distribution plot `fairscoring.plots.plot_groupwise_cdfs` and the difference `fairscoring.plots.plot plot_cdf_diffs`
+  to visualize independence, equal opportunity and predictive equality bias.
+- Custom colormaps in `fairscoring.plots.colors`
+
+#### Bias Result Objects
+- New result type `IntegralBiasResult` now contain the cdfs of the group distributions.
+
+### Changed
+- Submodules of `fairscoring.metrics` are now public.
+- The class `TwoGroupMixin` is now called `TwoGroupMetric`. 
+  This allows for a more consistent naming scheme.
+
 ## [0.1.1] - 2024-06-19
  
 ### Added